refactor: creates folders for the layers corresponding to the mcv architecture
